tests/fishery-pact/requirements.txt
pact-python==1.7.0
pytest==7.4.3
requests==2.31.0 